requirements-lock-cu130.txt
# CUDA 13 Windows lock profile for manga-to-realistic.
# PyTorch 2.10.0 is installed separately from the cu130 index by Install.bat.

accelerate==1.12.0
transformers==4.57.3
sentencepiece==0.2.1
protobuf==6.33.1
safetensors==0.6.2
huggingface_hub[hf_xet]==0.36.2
diffusers @ git+https://github.com/huggingface/diffusers@14f5ef3c0efa36dc6da8a44f1cdc893f10dc52f5
sdnq @ git+https://github.com/Disty0/sdnq.git@feb9659b130d7409a96b659347a19a28fe212c5f
gradio==6.5.1
pillow==10.4.0
peft==0.17.0
scipy==1.16.3
optimum-quanto==0.2.7
pruna==0.3.2; python_version < "3.13"
requests==2.32.5
nltk==3.9.2

opencv-python==4.13.0.90
timm==1.0.22
einops==0.8.1
ftfy==6.3.1
facexlib==0.3.0
insightface==0.7.3
scikit-learn==1.7.2
filterpy==1.4.5

controlnet-aux==0.0.10
# MediaPipe 0.10.32 is the latest version but removed the 'solutions' API
# ControlNet Aux will be degraded but the app handles this gracefully
mediapipe==0.10.32
qwen-tts==0.1.1
pyannote-audio==3.4.0
speechbrain==1.0.3
pydub==0.25.1
librosa==0.11.0
imageio-ffmpeg==0.6.0
soundfile==0.13.1
moviepy==2.1.2

# Optional:
# whisperx is intentionally excluded from this CUDA 13 lock profile because it
# currently pins torch~=2.8 and conflicts with PyTorch 2.10.0 cu130 wheels.
